refactor : 테스트 케이스가 순서를 가지지 않도록 수정한다
