+++ src/test/java/org/prgrms/kdt/TestConfiguration.java
@@ -1,5 +1,7 @@
 package org.prgrms.kdt;
 
+import com.wix.mysql.EmbeddedMysql;
+import com.wix.mysql.distribution.Version;
 import com.zaxxer.hikari.HikariDataSource;
 import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
@@ -8,6 +10,11 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
 import javax.sql.DataSource;
+
+import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
+import static com.wix.mysql.ScriptResolver.classPathScript;
+import static com.wix.mysql.config.Charset.UTF8;
+import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
 
 @Configuration
 @ComponentScan(
@@ -28,5 +35,17 @@ public class TestConfiguration {
     @Bean
     public NamedParameterJdbcTemplate jdbcTemplate(DataSource dataSource) {
         return new NamedParameterJdbcTemplate(dataSource);
+    }
+
+    public static void clean(EmbeddedMysql embeddedMysql) {
+        var mysqldConfig = aMysqldConfig(Version.v8_0_11)
+                .withCharset(UTF8)
+                .withPort(2215)
+                .withUser("test", "test1234!")
+                .withTimeZone("Asia/Seoul")
+                .build();
+        embeddedMysql = anEmbeddedMysql(mysqldConfig)
+                .addSchema("test-order-mgmt", classPathScript("schema.sql"))
+                .start();
     }
 }
