+++ src/test/java/org/prgrms/kdt/service/CustomerServiceTest.java
@@ -1,8 +1,8 @@
 package org.prgrms.kdt.service;
 
 import com.wix.mysql.EmbeddedMysql;
-import com.wix.mysql.distribution.Version;
 import org.junit.jupiter.api.*;
+import org.prgrms.kdt.TestConfiguration;
 import org.prgrms.kdt.model.customer.Customer;
 import org.prgrms.kdt.repository.CustomerJdbcRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +11,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
-import static com.wix.mysql.ScriptResolver.classPathScript;
-import static com.wix.mysql.config.Charset.UTF8;
-import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
@@ -36,16 +32,8 @@ class CustomerServiceTest {
 
     @BeforeAll
     void clean() {
+        TestConfiguration.clean(embeddedMysql);
         newCustomer = new Customer(UUID.randomUUID(), "test", "test@gmail.com", LocalDateTime.now(), LocalDateTime.now());
-        var mysqldConfig = aMysqldConfig(Version.v8_0_11)
-                .withCharset(UTF8)
-                .withPort(2215)
-                .withUser("test", "test1234!")
-                .withTimeZone("Asia/Seoul")
-                .build();
-        embeddedMysql = anEmbeddedMysql(mysqldConfig)
-                .addSchema("test-order-mgmt", classPathScript("schema.sql"))
-                .start();
         customerService.createCustomer(newCustomer);
     }
