+++ src/test/java/org/prgrms/kdt/service/VoucherServiceTest.java
@@ -1,13 +1,14 @@
 package org.prgrms.kdt.service;
 
 import com.wix.mysql.EmbeddedMysql;
-import com.wix.mysql.distribution.Version;
 import org.junit.jupiter.api.*;
+import org.prgrms.kdt.TestConfiguration;
 import org.prgrms.kdt.model.customer.Customer;
 import org.prgrms.kdt.model.voucher.FixedAmountVoucher;
 import org.prgrms.kdt.model.voucher.PercentDiscountVoucher;
 import org.prgrms.kdt.model.voucher.Voucher;
-import org.prgrms.kdt.repository.*;
+import org.prgrms.kdt.repository.CustomerJdbcRepository;
+import org.prgrms.kdt.repository.JdbcWalletRepository;
 import org.prgrms.kdt.repository.voucher.JdbcVoucherRepository;
 import org.prgrms.kdt.repository.voucher.MemoryVoucherRepository;
 import org.prgrms.kdt.repository.voucher.VoucherRepository;
@@ -20,10 +21,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
-import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
-import static com.wix.mysql.ScriptResolver.classPathScript;
-import static com.wix.mysql.config.Charset.UTF8;
-import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -51,15 +48,7 @@ class VoucherServiceTest {
 
     @BeforeAll
     void clean() {
-        var mysqldConfig = aMysqldConfig(Version.v8_0_11)
-                .withCharset(UTF8)
-                .withPort(2215)
-                .withUser("test", "test1234!")
-                .withTimeZone("Asia/Seoul")
-                .build();
-        embeddedMysql = anEmbeddedMysql(mysqldConfig)
-                .addSchema("test-order-mgmt", classPathScript("schema.sql"))
-                .start();
+        TestConfiguration.clean(embeddedMysql);
     }
 
     @AfterEach
@@ -189,7 +178,6 @@ class VoucherServiceTest {
     @DisplayName("고객에게 할당된 바우처를 삭제한다.")
     @Test
     void deleteVoucher() {
-        //VoucherService voucherService = mock(VoucherService.class);
         Customer customer = new Customer(UUID.randomUUID(), "test", "test@gmail.com", LocalDateTime.now(), LocalDateTime.now());
         Voucher voucher = new FixedAmountVoucher(UUID.randomUUID(), 300, LocalDateTime.now());
         customerRepository.insertCustomer(customer);
